+++ src/SettingsState.h
@@ -40,6 +40,7 @@ private:
   EnumValue* m_wait_tolerance_val;
   EnumValue* m_show_labels_val;
   EnumValue* m_guide_notes_val;
+  EnumValue* m_layout_val;
 
   // Tiles
   EnumTile* m_lead_in_tile;
@@ -50,6 +51,7 @@ private:
   EnumTile* m_wait_tolerance_tile;
   EnumTile* m_show_labels_tile;
   EnumTile* m_guide_notes_tile;
+  EnumTile* m_layout_tile;
 
   ButtonState m_test_audio_button;
 };